rag-orchestrator/requirements.txt
# --- Framework Principal ---
streamlit

# --- Conexión a Base de Datos Vectorial ---
chromadb==1.0.12

# --- Modelos de IA (Patrón 3) ---
sentence-transformers
transformers==4.36.2
torch
torchvision
torchaudio
# Para el modelo BLIP, a veces es necesario instalar desde git si hay problemas
# pip install salesforce-lavis

# --- Utilidades ---
requests
Pillow
python-dotenv
